requirements.txt
PyYAML>=6.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
numpy>=1.24
openai>=1.0.0
python-dotenv==1.0.0
pytz==2023.3
sendgrid==6.10.0
tqdm==4.65.0
click>=8.0.0
GitPython>=3.1.0
